Move variable category logic into the factory class
This pulls VariableCategoryFactory out into its own file and puts the
logic for rebuilding a variable category into the factory. Also adds
some cleanup logic so the variable observer will go away if the category
stops being used.
